RD-T39 Moving tests
