+++ src/__tests__/app/contacts.test.tsx
@@ -5,7 +5,7 @@ import { RefreshControl } from 'react-native';
 
 import { ContactType } from '@/models/v4/contacts/contactResultData';
 
-import Contacts from '../contacts';
+import Contacts from '../../app/(app)/contacts';
 
 // Mock dependencies
 jest.mock('react-i18next', () => ({
