+++ src/__tests__/app/signalr-lifecycle.test.tsx
